==> 1.Data_preprocessing-checkpoint.ipynb <==
import pandas as pd
import numpy as np

#############################################
######## Hornet Data Train - preprocessing###
#############################################
hornet_dataset_train = pd.read_excel(r'HORNET_PressureTags_2017_04_15_to_2017_06_01.xlsx')
hornet_data_upper_annular = hornet_dataset_train[['last_active_time','Active','upper_ann_open_current','upper_ann_close_current','UA RegPilotPes','UA RegRdbakPes','pod_pilot_pres_podtrans','pod_supply_pres_podtrans', 'pod_1_adj', 'pod_2_adj']]

hornet_dataset_train.head()

hornet_dataset_train.describe()

list(hornet_dataset_train.columns.values)

list(hornet_data_upper_annular.columns.values)

hornet_data_upper_annular["UA RegPilotPes"].head()

### Pressure Adjustments ###
hornet_data_upper_annular['UA RegPilotPes Adj'] = np.where(hornet_data_upper_annular['Active'] == 1, hornet_data_upper_annular['UA RegPilotPes'] - hornet_data_upper_annular['pod_1_adj'], hornet_data_upper_annular['UA RegPilotPes'] - hornet_data_upper_annular['pod_2_adj'])

#Logic of elsif statement i.e. where(condition, value1, value2)
hornet_data_upper_annular['UA RegRdbakPes Adj'] = np.where(hornet_data_upper_annular['Active'] == 1, hornet_data_upper_annular['UA RegRdbakPes'] - hornet_data_upper_annular['pod_1_adj'], hornet_data_upper_annular['UA RegRdbakPes'] - hornet_data_upper_annular['pod_2_adj'])

list(hornet_data_upper_annular.columns.values)

hornet_data_upper_annular['pod_pilot_pres_podtrans_adj'] = np.where(hornet_data_upper_annular['Active'] == 1, hornet_data_upper_annular['pod_pilot_pres_podtrans'] - hornet_data_upper_annular['pod_1_adj'], hornet_data_upper_annular['pod_pilot_pres_podtrans'] - hornet_data_upper_annular['pod_2_adj'])

hornet_data_upper_annular['pod_supply_pres_podtrans_adj'] = np.where(hornet_data_upper_annular['Active'] == 1, hornet_data_upper_annular['pod_supply_pres_podtrans'] - hornet_data_upper_annular['pod_1_adj'], hornet_data_upper_annular['pod_supply_pres_podtrans'] - hornet_data_upper_annular['pod_2_adj'])

hornet_data_upper_annular.to_csv(r'hornet_data_upper_annular_train.csv', index = False)

#############################################
######## Hornet Data Test - preprocessing####
#############################################
## Data in hornet_data_upper_annular_test 
## test data that is from 2nd of June till 20 August is for testing purpose 
## in the end of 1.pre_processing python, 
## you see hornet_data_upper_annular.to_csv(r'hornet_data_upper_annular_test.csv', index = False) 
hornet_dataset_test = pd.read_csv(r'HORNET_PressureTags_2017_06_02_to_2017_08_20_test_predicted.csv')

hornet_data_upper_annular = hornet_dataset_test[['last_active_time','Active','upper_ann_open_current','upper_ann_close_current','UA RegPilotPes','UA RegRdbakPes','pod_pilot_pres_podtrans','pod_supply_pres_podtrans', 'pod_1_adj', 'pod_2_adj']]

### Pressure Adjustments ###
hornet_data_upper_annular['UA RegPilotPes Adj'] = np.where(hornet_data_upper_annular['Active'] == 1, hornet_data_upper_annular['UA RegPilotPes'] - hornet_data_upper_annular['pod_1_adj'], hornet_data_upper_annular['UA RegPilotPes'] - hornet_data_upper_annular['pod_2_adj'])

hornet_data_upper_annular['UA RegRdbakPes Adj'] = np.where(hornet_data_upper_annular['Active'] == 1, hornet_data_upper_annular['UA RegRdbakPes'] - hornet_data_upper_annular['pod_1_adj'], hornet_data_upper_annular['UA RegRdbakPes'] - hornet_data_upper_annular['pod_2_adj'])

hornet_data_upper_annular['pod_pilot_pres_podtrans_adj'] = np.where(hornet_data_upper_annular['Active'] == 1, hornet_data_upper_annular['pod_pilot_pres_podtrans'] - hornet_data_upper_annular['pod_1_adj'], hornet_data_upper_annular['pod_pilot_pres_podtrans'] - hornet_data_upper_annular['pod_2_adj'])

hornet_data_upper_annular['pod_supply_pres_podtrans_adj'] = np.where(hornet_data_upper_annular['Active'] == 1, hornet_data_upper_annular['pod_supply_pres_podtrans'] - hornet_data_upper_annular['pod_1_adj'], hornet_data_upper_annular['pod_supply_pres_podtrans'] - hornet_data_upper_annular['pod_2_adj'])

hornet_data_upper_annular.to_csv(r'hornet_data_upper_annular_test.csv', index = False)
