==> 2_ModelA_clustering-checkpoint.ipynb <==
import pandas as pd
from sklearn.cluster import KMeans
import numpy as np
import pickle

#0 hornet_data_upper_annular_train = pd.read_csv(r'hornet_data_upper_annular_train.csv') 
#LION_2017_04_24_to_2017_08_07
lion_data_upper_annular_train = pd.read_csv(r'LION_2017_04_24_to_2017_08_07.csv') 

#should be using only train data instead of entire data
X_train = np.array(lion_data_upper_annular_train[['upper_ann_open_current', 'upper_ann_close_current', 'UA RegPilotPes_adj', 'UA RegRdbakPes_adj', 'pod_pilot_pres_podtrans_adj', 'pod_supply_pres_podtrans_adj']])

#>>>>>How was number of clusters concluded as 13 for Kmean model buidling.
kmeans_model = KMeans(n_clusters=13, random_state=0, max_iter = 600).fit(X_train)

kmeans_model.labels_

lion_data_upper_annular_train['predicted_state'] = kmeans_model.labels_

#>>>>What does 1,2,0 type anomaly means
#>>> how did you select certain labels as anomaly e.g. 2,4,8 
lion_data_upper_annular_train['anomaly'] = np.where(lion_data_upper_annular_train['predicted_state'].isin([2,4,8]),2,np.where(lion_data_upper_annular_train['predicted_state'].isin([5,6]),1,0))

lion_data_upper_annular_train.to_csv(r'lion_data_upper_annular_train_cluster1_trial_2.csv', index = False)

lion_data_upper_annular_train[['predicted_state']]

### Save Model

file = open(r'kmeans_model.pkl', 'wb')
pickle.dump(kmeans_model, file)
