==> 3_ModelB_sequence_detection-checkpoint.ipynb <==
import pandas as pd
import numpy as np
from sklearn.cluster import KMeans
import pickle
import re

lion_data_upper_annular_train = pd.read_csv(r'lion_data_upper_annular_train_cluster1.csv') 

#### Sequence of 4 scoring function definition
## >>> Why scoring function is required? Logic!
def seqof4_scoring(data):
    data = '_'.join(map(str, data))
    data = data + str('_')

    score_dataset = pd.DataFrame()

    for a in range(0, 3):
        for b in range(0, 3):
            for c in range(0, 3):
                for d in range(0, 3):
                    pattern1 = str('(?=(') + str(a) + str('_') + str(b) + str('_') + str(c) + str('_') + str('))')
                    pattern2 = str('(?=(') + str(a) + str('_') + str(b) + str('_') + str(c) + str('_') + str(d) + str('_') + str('))')
                    ptrn = str(a) + str('_') + str(b) + str('_') + str(c) + str('_') + str(d) + str('_')
                    freq_pattern1 = len(re.findall(pattern1, data))
                    freq_pattern2 = len(re.findall(pattern2, data))
                    if freq_pattern1 == 0:
                        prb = 0
                    else:
                        prb = freq_pattern2 / freq_pattern1
                    score_dataset = score_dataset.append(pd.DataFrame([[ptrn, prb]]))
                    print(ptrn)

    score_dataset.columns = ['seqof4_pattern', 'probability_score']
    return score_dataset

### Open current sequence of states detection model
X_train = np.array(lion_data_upper_annular_train['upper_ann_open_current'])

kmeans_model = KMeans(n_clusters=3, random_state=0, max_iter = 600).fit(X_train.reshape(-1,1))

hornet_data_upper_annular_train['open_current_states_kmean'] = kmeans_model.labels_

file = open(r'UA_open_current_seq_detection_model.pkl', 'wb')

pickle.dump(kmeans_model, file)

### Close current sequence of states detection model
X_train = np.array(hornet_data_upper_annular_train['upper_ann_close_current'])

kmeans_model = KMeans(n_clusters=3, random_state=0, max_iter = 600).fit(X_train.reshape(-1,1))

hornet_data_upper_annular_train['close_current_states_kmean'] = kmeans_model.labels_

file = open(r'UA_close_current_seq_detection_model.pkl', 'wb')
pickle.dump(kmeans_model, file)

hornet_data_upper_annular_train.to_csv(r'hornet_data_upper_annular_train.csv', index = False)

##### Sequence of 4 pattern, Open current
data = np.array(hornet_data_upper_annular_train['open_current_states_kmean'])
temp_df_open_current = seqof4_scoring(data)
temp_df_open_current.to_csv(r'seqof4_open_current_pattern_scores.csv', index = False)

##### Sequence of 4 pattern, Open current
data = np.array(hornet_data_upper_annular_train['close_current_states_kmean'])
temp_df_close_current = seqof4_scoring(data)
temp_df_close_current.to_csv(r'seqof4_close_current_pattern_scores.csv', index = False)
